requirements.txt
absl-py==1.4.0
amqp==5.2.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astropy==5.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.1.0
blis==0.7.9
branca==0.7.2
buildozer==1.5.0
cachetools==5.3.0
casa-formats-io==0.2.1
catalogue==2.0.8
celery==5.4.0
certifi==2023.5.7
cffi==1.16.0
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
confection==0.0.4
configparser==7.0.0
constantly==23.10.4
contourpy==1.0.7
cryptography==44.0.1
cycler==0.11.0
cymem==2.0.7
Cython==3.0.11
daphne==4.1.2
dask==2023.5.0
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==5.0.7
djangorestframework==3.15.2
docutils==0.21.2
echo==0.8.0
et-xmlfile==1.1.0
executing==1.2.0
fast-histogram==0.11
fastjsonschema==2.19.0
filelock==3.12.0
fiona==1.9.6
flatbuffers==23.5.9
folium==0.17.0
fonttools==4.39.4
fqdn==1.5.1
freetype-py==2.4.0
frozendict==2.4.6
fsspec==2023.5.0
gast==0.4.0
geopandas==0.14.4
glue-core==1.10.0
glue-vispy-viewers==1.0.7
glueviz==1.2.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.54.2
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
hdbscan==0.8.37
hsluv==5.0.3
html5lib==1.1
hyperlink==21.0.0
idna==3.4
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.6.0
incremental==24.7.2
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.12.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
kombu==5.3.7
kronos==0.2.2
langcodes==3.3.0
libclang==16.0.0
lightning-utilities==0.10.1
locket==1.0.0
lxml==5.3.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.1.0
moviepy==1.0.3
mpl-scatter-density==0.7
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.9
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathy==0.10.1
patsy==0.5.6
peewee==3.17.7
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.5.1
preshed==3.0.8
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
psycopg2==2.9.10
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pvextractor==0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pyerfa==2.0.0.3
Pygments==2.15.1
PyOpenGL==3.1.7
pyOpenSSL==25.0.0
pyparsing==3.0.9
pypiwin32==223
pyproj==3.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
radio-beam==0.3.4
redis==5.0.7
referencing==0.31.0
regex==2023.5.5
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
service-identity==24.2.0
sh==2.0.7
shapely==2.0.4
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spectral-cube==0.6.0
SQLAlchemy==2.0.32
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.14.2
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
Theano==1.0.5
thinc==8.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.3.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==24.11.0
txaio==23.1.1
typer==0.7.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
vine==5.1.0
virtualenv==20.23.0
vispy==0.13.0
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.4
whitenoise==6.6.0
wrapt==1.14.1
xlrd==2.0.1
xyzservices==2024.6.0
yfinance==0.2.48
zipp==3.15.0
zope.interface==7.2
